50 - Protecting the categories relations

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
 
@@ -89,6 +90,10 @@ class CategoryController extends Controller
   public function destroy(string $id)
   {
     $category = Category::findOrfail($id);
+    $subCategory = SubCategory::where('category_id', $category->id)->count();
+    if ($subCategory > 0) {
+      return response(['status' => 'error', 'message' => 'This items contrain, sub items for delete this you have to delete the sub items first!']);
+    }
     $category->delete();
     return response(['status' => 'success', 'Deleted Successfully!']);
   }
